Add support for VB enum to int conversions
Closes #180

+++ Tests/CSharp/StatementTests.cs
@@ -1403,7 +1403,7 @@ public class TestClass2
     {
         switch (true)
         {
-            case object _ when (DateAndTime.Today.DayOfWeek == DayOfWeek.Saturday) | (DateAndTime.Today.DayOfWeek == DayOfWeek.Sunday):
+            case object _ when ((int)DateAndTime.Today.DayOfWeek == (int)DayOfWeek.Saturday) | ((int)DateAndTime.Today.DayOfWeek == (int)DayOfWeek.Sunday):
                 {
                     // we do not work on weekends
                     return false;
